requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Flask-Cors==4.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
gunicorn==21.2.0
